refactor: Add user class
