+++ lb/Model/GameContent.swift
@@ -1,0 +1,13 @@
+//
+//  GameContent.swift
+//  lb
+//
+//  Created by Mac-HOME on 04.01.2020.
+//  Copyright © 2020 Mac-HOME. All rights reserved.
+//
+
+import Foundation
+
+struct GameContent {
+    
+}
