+++ lb/Resources/lose.png
