+++ lb/ViewModel/Game/GameViewModel.swift
@@ -1,0 +1,25 @@
+//
+//  GameViewModel.swift
+//  lb
+//
+//  Created by Mac-HOME on 04.01.2020.
+//  Copyright © 2020 Mac-HOME. All rights reserved.
+//
+
+import Firebase
+
+class GameViewModel {
+    
+    
+    
+    // Save user score
+    func saveScore() {
+        
+        if isCreator {
+            self.db.collection("games").document(gameID).setData(["creator_score": userCorrectAnswersCount], merge: true)
+        }
+        else {
+            self.db.collection("games").document(gameID).setData(["opponent_score": userCorrectAnswersCount], merge: true)
+        }
+    }
+}
